requirements.txt
Flask==2.0.2
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2
zipp==3.6.0
